% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{Weather Condition of Flagstaff, AZ , from 1950 to 2019.}
\format{
a data frame with 25449 observations
\describe{
  \item{Date}{The date of observation as a POSIX date format.}
  \item{PRCP}{Precipitation.}
  \item{SNOWBALL}{Snowfall.}
  \item{TMIN}{The minimum temperature.}
  \item{TMAX}{The maximum temperature.}
}
}
\source{
\url{www.ncdc.noaa.gov}
}
\usage{
Flagstaff_Weather
}
\description{
Weather Condition of Flagstaff, AZ , from 1950 to 2019.
}
\keyword{datasets}
